Disable SC2120, and SC2034
Shellcheck has 2 "issues" with the call to pglet_wait_event from line 232.

Both of the issues relate simply to shellcheck trying to assume the line
232 call is the only way pglet_wait_event will be called. That is not the
case here both in terms of the lack of a parameter being passed in nor the
PGLET_EVENT_DATA being unused.

There is nothing to see here and therefore we simple disable the errors.